# docs/10_api_openapi.yaml
openapi: 3.0.3
info:
  title: Freight Forwarding and Finance Platform API
  version: 2.0.0
  description: >-
    Organization- and branch-scoped API for freight operations, informational
    service charges, reusable financial documents, and double-entry accounting.
servers:
  - url: /api/v1
    description: Current API
security:
  - bearerAuth: []
...
paths:
...
  /quotations:
    get:
      tags: [Quotations]
      operationId: listQuotations
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}, {$ref: '#/components/parameters/BranchFilter'}, {name: status, in: query, schema: {type: string}}, {name: customer_party_id, in: query, schema: {type: integer}}]
      responses: {'200': {description: Quotations, content: {application/json: {schema: {$ref: '#/components/schemas/PagedQuotationResponse'}}}}}
    post:
      tags: [Quotations]
      operationId: createQuotation
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/QuotationCreate'}}}
      responses: {'201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Quotation'}}}}}
  /quotations/{quotationId}/revisions:
    post:
      tags: [Quotations]
      operationId: createQuotationRevision
      parameters: [{$ref: '#/components/parameters/QuotationId'}]
      requestBody:
        content: {application/json: {schema: {$ref: '#/components/schemas/RevisionCreate'}}}
      responses: {'201': {description: Revision created}}
  /quotation-revisions/{revisionId}/send:
    post:
      tags: [Quotations]
      operationId: sendQuotationRevision
      parameters: [{$ref: '#/components/parameters/RevisionId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {description: Revision sent}}
  /quotation-revisions/{revisionId}/accept:
    post:
      tags: [Quotations]
      operationId: acceptQuotationRevision
      parameters: [{$ref: '#/components/parameters/RevisionId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {description: Revision accepted}}
  /quotation-revisions/{revisionId}/convert:
    post:
      tags: [Quotations]
      operationId: convertQuotationRevision
      parameters: [{$ref: '#/components/parameters/RevisionId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'201': {description: Service order created}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    RevisionId: {name: revisionId, in: path, required: true, schema: {type: integer, format: int64}}
...
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
    BranchFilter: {name: branch_id, in: query, schema: {type: integer, format: int64}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 128}}
...
  schemas:
...
    PageMeta:
      type: object
      properties: {page: {type: integer}, page_size: {type: integer}, total: {type: integer}}
...
    QuotationCreate:
      type: object
      required: [branch_id, customer_party_id, trade_direction_id, currency_code, revision]
      properties: {branch_id: {type: integer}, customer_party_id: {type: integer}, trade_direction_id: {type: integer}, currency_code: {type: string, minLength: 3, maxLength: 3}, revision: {$ref: '#/components/schemas/RevisionCreate'}}
    RevisionCreate:
      type: object
      required: [quotation_date, currency_code]
      properties: {quotation_date: {type: string, format: date}, valid_until: {type: string, format: date}, currency_code: {type: string, minLength: 3, maxLength: 3}, description: {type: string}, notes: {type: string}, places: {type: array, items: {type: object}}, containers: {type: array, items: {type: object}}, lines: {type: array, items: {type: object}}}
...
    Quotation: {type: object, properties: {id: {type: integer}, quotation_no: {type: string}, organization_id: {type: integer}, branch_id: {type: integer}, status: {type: string}}}
    PagedQuotationResponse: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Quotation'}}, meta: {$ref: '#/components/schemas/PageMeta'}}}
...